Create bidd
work bidding NFT invoice integration ERC20 smart contract using Solidity: This smart contract allows the owner to create an ERC20 token that represents an invoice. Bidders can place bids on the invoice by transferring the token to the smart contract. The owner can then pay the invoice by calling the pay function, which distributes the bids to the bidders and takes a commission. The minimum bid amount and deadline can be set when creating the token. This smart contract can be modified to integrate NFTs by storing the invoice data in the NFT's metadata.